refactor(token-smuggling): modularize converters and update tests
