+++ pyrit/prompt_converter/__init__.py
@@ -8,7 +8,12 @@ from pyrit.prompt_converter.add_image_to_video_converter import AddImageVideoCon
 from pyrit.prompt_converter.add_text_image_converter import AddTextImageConverter
 from pyrit.prompt_converter.ansi_escape.ansi_attack_converter import AnsiAttackConverter
 from pyrit.prompt_converter.ascii_art_converter import AsciiArtConverter
-from pyrit.prompt_converter.ascii_smuggler_converter import AsciiSmugglerConverter
+from pyrit.prompt_converter.token_smuggling import (
+    SmugglerConverter,
+    AsciiSmugglerConverter,
+    SneakyBitsSmuggler,
+    VariationSelectorSmuggler,
+)
 from pyrit.prompt_converter.atbash_converter import AtbashConverter
 from pyrit.prompt_converter.audio_frequency_converter import AudioFrequencyConverter
 from pyrit.prompt_converter.azure_speech_audio_to_text_converter import AzureSpeechAudioToTextConverter
@@ -100,6 +105,8 @@ __all__ = [
     "RepeatTokenConverter",
     "ROT13Converter",
     "SearchReplaceConverter",
+    "SmugglerConverter",
+    "SneakyBitsSmuggler",
     "StringJoinConverter",
     "SuffixAppendConverter",
     "TextToHexConverter",
@@ -111,4 +118,5 @@ __all__ = [
     "UnicodeSubstitutionConverter",
     "UrlConverter",
     "VariationConverter",
+    "VariationSelectorSmuggler",
 ]
