+++ tests/unit/converter/test_token_smuggler_converter.py
@@ -4,13 +4,17 @@
 import pytest
 
 from pyrit.prompt_converter import ConverterResult
-from pyrit.prompt_converter.ascii_smuggler_converter import AsciiSmugglerConverter
+from pyrit.prompt_converter.token_smuggling import (
+    AsciiSmugglerConverter,
+    SneakyBitsSmuggler,
+    VariationSelectorSmuggler,
+)
 
 
 @pytest.mark.asyncio
 async def test_convert_async_encode_unicode_tags_control():
     # Test encoding using the Unicode Tags mode with control tags enabled.
-    converter = AsciiSmugglerConverter(action="encode", encoding_mode="unicode_tags_control")
+    converter = AsciiSmugglerConverter(action="encode", unicode_tags=True)
     prompt = "Hello, World!"
     result = await converter.convert_async(prompt=prompt, input_type="text")
     assert isinstance(result, ConverterResult)
@@ -23,7 +27,7 @@ async def test_convert_async_encode_unicode_tags_control():
 @pytest.mark.asyncio
 async def test_convert_async_encode_unicode_tags():
     # Test encoding using the Unicode Tags mode without control tags.
-    converter = AsciiSmugglerConverter(action="encode", encoding_mode="unicode_tags")
+    converter = AsciiSmugglerConverter(action="encode", unicode_tags=False)
     prompt = "Hello, World!"
     result = await converter.convert_async(prompt=prompt, input_type="text")
     assert isinstance(result, ConverterResult)
@@ -36,7 +40,7 @@ async def test_convert_async_encode_unicode_tags():
 async def test_convert_async_decode_unicode_tags():
     # The following encoded message is "Hi" with control tags
     encoded_message = chr(0xE0001) + chr(0xE0000) + chr(0xE0048) + chr(0xE0000) + chr(0xE0069) + chr(0xE007F)
-    converter = AsciiSmugglerConverter(action="decode", encoding_mode="unicode_tags_control")
+    converter = AsciiSmugglerConverter(action="decode", unicode_tags=True)
     result = await converter.convert_async(prompt=encoded_message, input_type="text")
     assert isinstance(result, ConverterResult)
     assert result.output_type == "text"
@@ -47,10 +51,10 @@ async def test_convert_async_decode_unicode_tags():
 async def test_encode_decode_unicode_tags_control():
     # Test round-trip encoding/decoding with Unicode Tags control mode.
     base_string = "Hello, World!"
-    encode_converter = AsciiSmugglerConverter(action="encode", encoding_mode="unicode_tags_control")
+    encode_converter = AsciiSmugglerConverter(action="encode", unicode_tags=True)
     encoded_result = await encode_converter.convert_async(prompt=base_string, input_type="text")
 
-    decode_converter = AsciiSmugglerConverter(action="decode", encoding_mode="unicode_tags_control")
+    decode_converter = AsciiSmugglerConverter(action="decode", unicode_tags=True)
     decoded_result = await decode_converter.convert_async(prompt=encoded_result.output_text, input_type="text")
 
     assert isinstance(decoded_result, ConverterResult)
@@ -62,10 +66,10 @@ async def test_encode_decode_unicode_tags_control():
 async def test_encode_decode_unicode_tags():
     # Test encoding and decoding without unicode tags to ensure input integrity.
     base_string = "Hello, World!"
-    encode_converter = AsciiSmugglerConverter(action="encode", encoding_mode="unicode_tags")
+    encode_converter = AsciiSmugglerConverter(action="encode", unicode_tags=False)
     encoded_result = await encode_converter.convert_async(prompt=base_string, input_type="text")
 
-    decode_converter = AsciiSmugglerConverter(action="decode", encoding_mode="unicode_tags")
+    decode_converter = AsciiSmugglerConverter(action="decode", unicode_tags=False)
     decoded_result = await decode_converter.convert_async(prompt=encoded_result.output_text, input_type="text")
 
     assert isinstance(decoded_result, ConverterResult)
@@ -76,7 +80,7 @@ async def test_encode_decode_unicode_tags():
 @pytest.mark.asyncio
 async def test_convert_async_encode_sneaky_bits():
     # Test encoding using the Sneaky Bits mode.
-    converter = AsciiSmugglerConverter(action="encode", encoding_mode="sneaky_bits")
+    converter = SneakyBitsSmuggler(action="encode")
     prompt = "Hello, World!"
     result = await converter.convert_async(prompt=prompt, input_type="text")
     assert isinstance(result, ConverterResult)
@@ -90,10 +94,10 @@ async def test_convert_async_encode_sneaky_bits():
 async def test_convert_async_decode_sneaky_bits():
     # Test decoding using the Sneaky Bits mode.
     original_text = "Test Sneaky Bits"
-    encode_converter = AsciiSmugglerConverter(action="encode", encoding_mode="sneaky_bits")
+    encode_converter = SneakyBitsSmuggler(action="encode")
     encoded_result = await encode_converter.convert_async(prompt=original_text, input_type="text")
 
-    decode_converter = AsciiSmugglerConverter(action="decode", encoding_mode="sneaky_bits")
+    decode_converter = SneakyBitsSmuggler(action="decode")
     decoded_result = await decode_converter.convert_async(prompt=encoded_result.output_text, input_type="text")
     assert isinstance(decoded_result, ConverterResult)
     assert decoded_result.output_type == "text"
@@ -105,7 +109,7 @@ async def test_convert_async_encode_only_modes():
     prompt = "Test"
 
     # Unicode Tags Control
-    converter_control = AsciiSmugglerConverter(action="encode", encoding_mode="unicode_tags_control")
+    converter_control = AsciiSmugglerConverter(action="encode", unicode_tags=True)
     result_control = await converter_control.convert_async(prompt=prompt, input_type="text")
     assert result_control.output_text != ""
     assert result_control.output_text.startswith(chr(0xE0001))
@@ -113,7 +117,7 @@ async def test_convert_async_encode_only_modes():
     assert len(result_control.output_text) == len(prompt) + 2  # Explicit length check
 
     # Unicode Tags (no control)
-    converter_no_control = AsciiSmugglerConverter(action="encode", encoding_mode="unicode_tags")
+    converter_no_control = AsciiSmugglerConverter(action="encode", unicode_tags=False)
     result_no_control = await converter_no_control.convert_async(prompt=prompt, input_type="text")
     assert result_no_control.output_text != ""
     assert not result_no_control.output_text.startswith(chr(0xE0001))
@@ -121,7 +125,7 @@ async def test_convert_async_encode_only_modes():
     assert len(result_no_control.output_text) == len(prompt)  # Explicit length check
 
     # Sneaky Bits
-    converter_sneaky = AsciiSmugglerConverter(action="encode", encoding_mode="sneaky_bits")
+    converter_sneaky = SneakyBitsSmuggler(action="encode")
     result_sneaky = await converter_sneaky.convert_async(prompt=prompt, input_type="text")
     assert result_sneaky.output_text != ""
     valid_chars = {converter_sneaky.zero_char, converter_sneaky.one_char}
@@ -132,7 +136,7 @@ async def test_convert_async_encode_only_modes():
 @pytest.mark.asyncio
 async def test_convert_async_encode_variation_selector_smuggler():
     # Test encoding using the variation_selector_smuggler mode.
-    converter = AsciiSmugglerConverter(action="encode", encoding_mode="variation_selector_smuggler", embed_in_base=True)
+    converter = VariationSelectorSmuggler(action="encode", embed_in_base=True)
     prompt = "Hello, World!"
     result = await converter.convert_async(prompt=prompt, input_type="text")
     assert isinstance(result, ConverterResult)
@@ -154,15 +158,10 @@ async def test_convert_async_decode_variation_selector_smuggler():
     # Test decoding using the variation_selector_smuggler mode.
     # First encode a known string.
     original_text = "Hello, World!"
-    encode_converter = AsciiSmugglerConverter(
-        action="encode", encoding_mode="variation_selector_smuggler", embed_in_base=True
-    )
+    encode_converter = VariationSelectorSmuggler(action="encode", embed_in_base=True)
     encoded_result = await encode_converter.convert_async(prompt=original_text, input_type="text")
 
-    # Now decode the encoded result.
-    decode_converter = AsciiSmugglerConverter(
-        action="decode", encoding_mode="variation_selector_smuggler", embed_in_base=True
-    )
+    decode_converter = VariationSelectorSmuggler(action="decode", embed_in_base=True)
     decoded_result = await decode_converter.convert_async(prompt=encoded_result.output_text, input_type="text")
 
     assert isinstance(decoded_result, ConverterResult)
@@ -174,14 +173,10 @@ async def test_convert_async_decode_variation_selector_smuggler():
 async def test_encode_decode_variation_selector_smuggler_multibyte():
     # Test round-trip encoding/decoding with multibyte characters.
     base_string = "Ciao, mondo! 😊"
-    encode_converter = AsciiSmugglerConverter(
-        action="encode", encoding_mode="variation_selector_smuggler", embed_in_base=True
-    )
+    encode_converter = VariationSelectorSmuggler(action="encode", embed_in_base=True)
     encoded_result = await encode_converter.convert_async(prompt=base_string, input_type="text")
 
-    decode_converter = AsciiSmugglerConverter(
-        action="decode", encoding_mode="variation_selector_smuggler", embed_in_base=True
-    )
+    decode_converter = VariationSelectorSmuggler(action="decode", embed_in_base=True)
     decoded_result = await decode_converter.convert_async(prompt=encoded_result.output_text, input_type="text")
 
     assert isinstance(decoded_result, ConverterResult)
@@ -191,8 +186,8 @@ async def test_encode_decode_variation_selector_smuggler_multibyte():
 
 @pytest.mark.asyncio
 async def test_encode_decode_visible_hidden():
-    converter = AsciiSmugglerConverter(encoding_mode="variation_selector_smuggler", embed_in_base=True)
-
+    # Test mixing visible + hidden text
+    converter = VariationSelectorSmuggler(embed_in_base=True)
     visible_text = "Hallo wie geht es dir?"
     hidden_text = "Das ist eine geheime Nachricht!"
 
@@ -212,12 +207,10 @@ async def test_encode_decode_visible_hidden():
 
 @pytest.mark.asyncio
 async def test_embed_in_base_false_inserts_separator():
-    # Test that when embed_in_base is False.
-    converter = AsciiSmugglerConverter(
-        action="encode", encoding_mode="variation_selector_smuggler", embed_in_base=False
-    )
+    # Test that when embed_in_base is False, space is inserted after base char.
+    converter = VariationSelectorSmuggler(action="encode", embed_in_base=False)
     prompt = "Secret"
-    _, encoded = converter._encode_variation_selector_smuggler(prompt)
+    _, encoded = converter.encode_message(prompt)
     base_char = converter.utf8_base_char
     # Expect the encoded string to be: base_char + " " + payload
     assert encoded.startswith(
@@ -227,6 +220,7 @@ async def test_embed_in_base_false_inserts_separator():
 
 # Test for the input_supported method
 def test_input_supported():
+    # Verify input type filtering
     converter = AsciiSmugglerConverter()
     assert converter.input_supported("text") is True
     assert converter.input_supported("image") is False
